GME.sln 64 bit compilation
git-svn-id: https://svn.isis.vanderbilt.edu/GMESRC/trunk@1609 b35830b6-564f-0410-ac5f-b387e50cb686

+++ GME/Console/ConsoleCtl.h
@@ -106,7 +106,7 @@ public:
 	eventidClickMGAID = 1L,
 	//}}AFX_DISP_ID
 	};
-	afx_msg int GetCWnd() { return (int)(void*)(CWnd*)this; }
+	afx_msg __int64 GetCWnd() { return (__int64)(void*)(CWnd*)this; }
 	afx_msg void SetCWnd(void*) {  }
 	BOOL PreTranslateMessage(MSG* pMsg);
 };
